hotshot_example_types
Module node_types

node types

Structs

CombinedImpl

Combined Network implementation (libp2p + web server)

EpochUpgradeTestVersions

EpochsTestVersions

Libp2pImpl

Libp2p network implementation

MarketplaceTestVersions

MarketplaceUpgradeTestVersions

MemoryImpl

Memory network implementation

PushCdnImpl

The Push CDN implementation

RandomOverlapQuorumFilterConfig

Provides parameters to use the RandomOverlapQuorumIterator

StableQuorumFilterConfig

Provides parameters to use the StableQuorumIterator

TestConsecutiveLeaderTypes

filler struct to implement node type and allow us to select our traits

TestTwoStakeTablesTypes

filler struct to implement node type and allow us to select our traits

TestTypes

filler struct to implement node type and allow us to select our traits

TestTypesRandomizedCommitteeMembers

filler struct to implement node type and allow us to select our traits

TestTypesRandomizedLeader

filler struct to implement node type and allow us to select our traits

TestVersions

WebImpl

Web server network implementation

Type Aliases

StaticMembership

static committee type alias