Module hotshot_example_types::testable_delay

add a delay to async functions

Structs

DelayConfig

Settings for each type

DelaySettings

Config for each supported type

SupportedTraitTypesForAsyncDelayIterator 🔒

Iterator to iterate over enum

Enums

DelayOptions

What type of delay we want to apply to

SupportedTraitTypesForAsyncDelay

Current implementations that are supported for testing async delays

Traits

TestableDelay

Implement this method to add some delay to async call