Trait hotshot_example_types::testable_delay::TestableDelay

pub trait TestableDelay {
    // Required method
    fn run_delay_settings_from_config<'life0, 'async_trait>(
        delay_config: &'life0 DelayConfig,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where 'life0: 'async_trait;

    // Provided method
    fn handle_async_delay<'life0, 'async_trait>(
        settings: &'life0 DelaySettings,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where 'life0: 'async_trait { ... }
}

Implement this method to add some delay to async call

Required Methods

fn run_delay_settings_from_config<'life0, 'async_trait>( delay_config: &'life0 DelayConfig, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where 'life0: 'async_trait,

Look for settings in the config and run it

Provided Methods

fn handle_async_delay<'life0, 'async_trait>( settings: &'life0 DelaySettings, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where 'life0: 'async_trait,

Add a delay from settings

Object Safety

This trait is not object safe.